reorganized common, updated related imports
