+++ common/js/Board/__tests__/Board.test.js
@@ -1,11 +1,7 @@
 const Board = require('common/js/Board');
 const Piece = require('common/js/Piece');
-const {
-  PIECES, ROTATE_LEFT, ROTATE_RIGHT, MAX_FLOOR_KICKS,
-} = require('common/helpers/constants');
-const {
-  getTestBoard, getTestPiece, getTestPieces,
-} = require('common/mockData/mocks');
+const { PIECES, ROTATE_LEFT, ROTATE_RIGHT, MAX_FLOOR_KICKS } = require('common/constants');
+const { getTestBoard, getTestPiece, getTestPieces } = require('common/mocks');
 const pubSub = require('frontend/helpers/pubSub');
 
 const movePieceTo = (gameBoard, location) => {
