+++ common/js/Game/__tests__/Game.test.js
@@ -1,11 +1,9 @@
 const Game = require('common/js/Game');
 const Board = require('common/js/Board');
 const Piece = require('common/js/Piece');
-const { LINES_PER_LEVEL, POINTS } = require('common/helpers/constants');
-const {
-  ADD_PIECES, LOWER_PIECE, CLEAR_LINES, END_GAME,
-} = require('common/helpers/commonTopics');
-const { getTestBoard, getTestPiece, getTestPieces } = require('common/mockData/mocks');
+const { LINES_PER_LEVEL, POINTS } = require('common/constants');
+const { ADD_PIECES, LOWER_PIECE, CLEAR_LINES, END_GAME} = require('common/topics');
+const { getTestBoard, getTestPiece, getTestPieces } = require('common/mocks');
 const pubSub = require('backend/helpers/pubSub');
 
 describe('game tests', () => {
