+++ frontend/__tests__/ClientBoard.test.js
@@ -1,9 +1,7 @@
 const ClientBoard = require('frontend/static/js/ClientBoard');
 const { ROTATE_LEFT } = require('frontend/helpers/clientConstants');
 const { DRAW, CLEAR_LINES, BOARD_CHANGE } = require('frontend/helpers/clientTopics');
-const {
-  getTestBoard, getTestPiece, getTestPieces, pubSubMock,
-} = require('common/mockData/mocks');
+const { getTestBoard, getTestPiece, getTestPieces, pubSubMock } = require('common/mocks');
 const pubSub = require('frontend/helpers/pubSub');
 
 describe('client - game board tests', () => {
