+++ src/__tests__/GameManager.test.js
@@ -3,7 +3,7 @@ const MessageManager = require('backend/js/MessageManager');
 const PlayerManager = require('backend/js/PlayerManager');
 const { GAME_TYPES, COUNTDOWN, POWER_UP_TYPES } = require('backend/helpers/serverConstants');
 const { ADD_PIECES, END_GAME, MSG_TYPE } = require('backend/helpers/serverTopics');
-const { getTestBoard } = require('common/mockData/mocks');
+const { getTestBoard } = require('common/mocks');
 
 describe('game manager tests', () => {
   let gameManager;
