+++ src/__tests__/ServerGame.test.js
@@ -3,12 +3,7 @@ const { GAME_TYPES, POWER_UP_TYPES, MAX_POWER_UPS } = require('backend/helpers/s
 const {
   UPDATE_PLAYER, CLEAR_LINES, ADD_POWER_UP, USE_POWER_UP,
 } = require('backend/helpers/serverTopics');
-const {
-  getTestBoard,
-  getTestPiece,
-  getTestPieces,
-  pubSubMock,
-} = require('common/mockData/mocks');
+const { getTestBoard, getTestPiece, getTestPieces, pubSubMock } = require('common/mocks');
 const pubSub = require('backend/helpers/pubSub');
 
 describe('game tests', () => {
