+++ src/__tests__/routeHelpers.test.js
@@ -1,6 +1,6 @@
 const GameServer = require('backend/js/GameServer');
 const Player = require('backend/js/Player');
-const { GAME_TYPES } = require('common/helpers/constants');
+const { GAME_TYPES } = require('common/constants');
 const { PLAY, EXECUTE_COMMANDS } = require('backend/helpers/serverTopics');
 const {
   multiGameExists,
