+++ src/helpers/routeHelpers.js
@@ -1,4 +1,4 @@
-const { GAME_TYPES } = require('common/helpers/constants');
+const { GAME_TYPES } = require('common/constants');
 const { PLAY, EXECUTE_COMMANDS } = require('backend/helpers/serverTopics');
 const GameServer = require('backend/js/GameServer');
 const Player = require('backend/js/Player');
